refactor: MVC 구조를 위해 GameStatus와 OutputView의 의존관계 정리
GameStatus 기존에 OutputView를 호출해 로그를 출력하던 부분을, 메서드를 통해 반환하여 controller(Main)에서 OutputView로 전달하도록 수정하였음.
OutputView 필요없는 메서드 정리
Main GameStatus 구조 변경에 맞춰 로직 변경
RacingGame domain 역할에 충실하도록 로직 분리, moveCars 메서드 수정
